refactor: convert DashboardPanel to .tsx
