+++ frontend/src/components/common/dashboard/index.js
@@ -1,6 +1,5 @@
 export { default as MapComponent } from './MapComponent';
 export { default as MapFilterButton } from './MapFilterButton';
 export { default as MapLegend } from './MapLegend';
-export { default as DashboardPanel } from './DashboardPanel';
 export { default as TimeseriesComponent } from './TimeseriesComponent';
 export { default as TimeseriesControls } from './TimeseriesControls';
